requirements.txt
# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Orchestration ─────────────────────────────────────────────────────────────
prefect>=2.19,<3.0

# ── GCP SDKs ──────────────────────────────────────────────────────────────────
google-cloud-aiplatform>=1.55.0
google-cloud-storage>=2.16.0
google-cloud-firestore>=2.16.0

# ── Shared ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0
